requirements.txt
streamlit
pillow
numpy
google-generativeai
python-dotenv
opencv-python
ultralytics
torch
torchvision
transformers
requests
tqdm
google-cloud-storage
google-auth
google-auth-oauthlib
google-auth-httplib2
fastapi
uvicorn
requests
python-multipart